project_01_rag_agent/requirements.txt
langchain>=0.3
langchain-community>=0.3
langchain-core>=0.3
langchain-text-splitters>=0.3
langgraph>=0.2
langchain-chroma>=0.1
chromadb>=0.5
rank-bm25>=0.2
pypdf>=4.0
unstructured>=0.15
docx2txt>=0.8
fastapi>=0.115
uvicorn[standard]>=0.30
streamlit>=1.40
python-dotenv>=1.0
loguru>=0.7
httpx>=0.27
pydantic>=2.0
